refactor(core): split SDK error utils, run storage, auth session, skills loader
sdkErrorUtils becomes a barrel over sdkError/ (kinds, metadata,
inspection, relay detection, patterns, formatting); taskRunStorage,
SupabaseSession, and loadSkills keep their classes and re-export the
extracted value objects and helpers.
