+++ src/auth/SupabaseSession.ts
@@ -1,71 +1,46 @@
-import { z } from 'zod';
-import { isAbortError } from '@common/errors';
 import { toErrorMessage } from '@common/errors/errorMessage';
 import {
   parseAuthCallbackTokens,
   type AuthCallbackUriParts,
 } from './core/authCallback';
+import { fetchWithTimeout } from './fetchWithTimeout';
+import {
+  DEFAULT_SUPABASE_SESSION_EXPIRY_MS,
+  firstAccountLabel,
+  parseStoredSupabaseSession,
+  parseTokenExchangeResponse,
+  toStorableGitHubTokenExchangeSession,
+  toStorableSupabaseSession,
+  type SupabaseCallbackResult,
+  type SupabaseSession,
+  type SupabaseSessionLog,
+  type SupabaseSessionStorage,
+} from './supabaseSessionTypes';
 import type { AuthTokenProvider, SessionTokens } from './TokenProvider';
-import type {
-  Session as SupabaseNativeSession,
-  SupabaseClient as Client,
-} from '@supabase/supabase-js';
+import type { SupabaseClient as Client } from '@supabase/supabase-js';
 
-export const DEFAULT_SUPABASE_SESSION_EXPIRY_MS = 60 * 60 * 1000;
-
-export const SupabaseSessionSchema = z.object({
-  id: z.string(),
-  accessToken: z.string(),
-  refreshToken: z.string(),
-  account: z.object({
-    id: z.string(),
-    label: z.string().transform((label) => label.trim()),
-  }),
-  expiresAt: z.number(),
-  useCustomRefresh: z.boolean().optional(),
-});
-export type SupabaseSession = z.infer<typeof SupabaseSessionSchema>;
-
-/** Response schema for GitHub token exchange and refresh Edge Functions. */
-export const GitHubTokenExchangeSchema = z.object({
-  access_token: z.string(),
-  refresh_token: z.string(),
-  expires_at: z.number().optional(),
-  expires_in: z.number().optional(),
-  token_type: z.string(),
-  user: z.object({
-    id: z.string(),
-    email: z.string().nullish(),
-    user_metadata: z
-      .object({
-        avatar_url: z.string().optional(),
-        user_name: z.string().optional(),
-      })
-      .optional(),
-  }),
-});
-export type GitHubTokenExchangeResponse = z.infer<
-  typeof GitHubTokenExchangeSchema
->;
-
-export interface SupabaseSessionParseOptions {
-  logSource?: string;
-  warn?: (source: string, message: string) => void;
-}
-
-/** Storage boundary for persisted Supabase session data. */
-export interface SupabaseSessionStorage {
-  get(): Promise<string | undefined>;
-  store(sessionData: string): Promise<void>;
-  delete(): Promise<void>;
-}
-
-export interface SupabaseSessionLog {
-  debug?(source: string, message: string): void;
-  info?(source: string, message: string): void;
-  warn?(source: string, message: string): void;
-  error?(source: string, message: string): void;
-}
+// Public entry point: external consumers import session schemas, value-object
+// helpers, the fetch-with-timeout utility, and this coordinator through
+// `@auth/SupabaseSession`. Implementation is split into the modules above.
+export { fetchWithTimeout } from './fetchWithTimeout';
+export {
+  DEFAULT_SUPABASE_SESSION_EXPIRY_MS,
+  GitHubTokenExchangeSchema,
+  SupabaseSessionSchema,
+  firstAccountLabel,
+  parseStoredSupabaseSession,
+  parseTokenExchangeResponse,
+  toStorableGitHubTokenExchangeSession,
+  toStorableSupabaseSession,
+  type GitHubTokenExchangeResponse,
+  type SupabaseCallbackParseError,
+  type SupabaseCallbackParseResult,
+  type SupabaseCallbackResult,
+  type SupabaseSession,
+  type SupabaseSessionLog,
+  type SupabaseSessionParseOptions,
+  type SupabaseSessionStorage,
+} from './supabaseSessionTypes';
 
 export interface SupabaseSessionCoordinatorOptions {
   storage: SupabaseSessionStorage;
@@ -80,186 +55,9 @@ export interface SupabaseSessionCoordinatorOptions {
   onTokenExpiryChanged?: (expiresAt: number | null) => void;
 }
 
-/** Result of converting an auth callback into a stored session. */
-export interface SupabaseCallbackParseResult {
-  success: true;
-  session: SupabaseSession;
-}
-
-export interface SupabaseCallbackParseError {
-  success: false;
-  error: string;
-  isAuthError?: boolean;
-}
-
-export type SupabaseCallbackResult =
-  | SupabaseCallbackParseResult
-  | SupabaseCallbackParseError;
-
 interface StableSessionSnapshot {
   session: SupabaseSession | null;
   version: number;
-}
-
-function firstAccountLabel(
-  ...candidates: readonly (string | null | undefined)[]
-): string {
-  for (const candidate of candidates) {
-    const label = candidate?.trim();
-    if (label) return label;
-  }
-  return 'unknown';
-}
-
-/**
- * Parse and validate stored session data.
- * Returns null if session data is missing or invalid.
- * Logs warnings for corrupted data to help diagnose auth issues.
- */
-export function parseStoredSupabaseSession(
-  sessionData: string | undefined,
-  options?: SupabaseSessionParseOptions,
-): SupabaseSession | null {
-  if (!sessionData) return null;
-  const logSource = options?.logSource ?? 'SupabaseSession';
-  try {
-    const parsed = SupabaseSessionSchema.safeParse(JSON.parse(sessionData));
-    if (!parsed.success) {
-      options?.warn?.(
-        logSource,
-        `Stored session has invalid schema: ${parsed.error.message}`,
-      );
-      return null;
-    }
-    return parsed.data;
-  } catch (error) {
-    options?.warn?.(
-      logSource,
-      `Failed to parse stored session: ${toErrorMessage(error)}`,
-    );
-    return null;
-  }
-}
-
-export async function fetchWithTimeout(
-  url: string,
-  options: RequestInit,
-  timeoutMs: number,
-  timeoutMessage: string,
-  fetchImpl: typeof fetch = fetch,
-): Promise<Response> {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), timeoutMs);
-  const { signal, cleanup } = combineAbortSignals(
-    options.signal ?? undefined,
-    controller.signal,
-  );
-  try {
-    return await fetchImpl(url, { ...options, signal });
-  } catch (error) {
-    if (isAbortError(error)) {
-      if (options.signal?.aborted && !controller.signal.aborted) {
-        throw error;
-      }
-      throw new Error(timeoutMessage);
-    }
-    throw error;
-  } finally {
-    cleanup();
-    clearTimeout(timeoutId);
-  }
-}
-
-function combineAbortSignals(
-  upstreamSignal: AbortSignal | undefined,
-  timeoutSignal: AbortSignal,
-): { signal: AbortSignal; cleanup: () => void } {
-  if (!upstreamSignal) {
-    return { signal: timeoutSignal, cleanup: () => {} };
-  }
-
-  if (typeof AbortSignal.any === 'function') {
-    return {
-      signal: AbortSignal.any([upstreamSignal, timeoutSignal]),
-      cleanup: () => {},
-    };
-  }
-
-  const controller = new AbortController();
-  const abort = () => controller.abort();
-  upstreamSignal.addEventListener('abort', abort, { once: true });
-  timeoutSignal.addEventListener('abort', abort, { once: true });
-  if (upstreamSignal.aborted || timeoutSignal.aborted) {
-    controller.abort();
-  }
-
-  return {
-    signal: controller.signal,
-    cleanup: () => {
-      upstreamSignal.removeEventListener('abort', abort);
-      timeoutSignal.removeEventListener('abort', abort);
-    },
-  };
-}
-
-export async function parseTokenExchangeResponse(
-  response: Response,
-  log?: SupabaseSessionLog,
-): Promise<GitHubTokenExchangeResponse> {
-  const rawData = await response.json();
-  const parsed = GitHubTokenExchangeSchema.safeParse(rawData);
-  if (!parsed.success) {
-    log?.error?.(
-      'SupabaseSession',
-      `Token exchange response validation failed: ${parsed.error.message}`,
-    );
-    throw new Error('Invalid response format from authentication server');
-  }
-  return parsed.data;
-}
-
-/**
- * Convert Supabase's native Session to our storage format.
- * Handles the snake_case to camelCase and seconds to milliseconds conversions.
- */
-export function toStorableSupabaseSession(
-  nativeSession: SupabaseNativeSession,
-  options?: { useCustomRefresh?: boolean },
-): SupabaseSession {
-  return {
-    id: nativeSession.user.id,
-    accessToken: nativeSession.access_token,
-    refreshToken: nativeSession.refresh_token,
-    account: {
-      id: nativeSession.user.id,
-      label: firstAccountLabel(nativeSession.user.email, nativeSession.user.id),
-    },
-    expiresAt: nativeSession.expires_at
-      ? nativeSession.expires_at * 1000
-      : Date.now() + DEFAULT_SUPABASE_SESSION_EXPIRY_MS,
-    useCustomRefresh: options?.useCustomRefresh,
-  };
-}
-
-/** Convert Edge Function token responses into the stored custom-refresh shape. */
-export function toStorableGitHubTokenExchangeSession(
-  data: GitHubTokenExchangeResponse,
-  fallbackLabel: string,
-  defaultSessionExpiryMs: number,
-): SupabaseSession {
-  return {
-    id: data.user.id,
-    accessToken: data.access_token,
-    refreshToken: data.refresh_token,
-    account: {
-      id: data.user.id,
-      label: firstAccountLabel(data.user.email, fallbackLabel, data.user.id),
-    },
-    expiresAt: data.expires_at
-      ? data.expires_at * 1000
-      : Date.now() + defaultSessionExpiryMs,
-    useCustomRefresh: true,
-  };
 }
 
 /**
