+++ src/skills/loadSkills.ts
@@ -2,51 +2,25 @@
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 
-// Third-party imports
-import { ZodError } from 'zod';
-
 // Local imports - common
 import {
   isFileNotFoundError,
   isNotADirectoryError,
   toErrorMessage,
 } from '@common/errors';
-import { isObject } from '@utils/core';
 
 // Local imports - skill parsing
-import { collapseWhitespace } from '@utils/text/stringUtils';
-import {
-  SKILL_DESCRIPTION_MAX_LENGTH,
-  SkillNameSchema,
-  SkillSchema,
-  type Skill,
-} from './SkillSchema';
-import { extractFrontmatter } from './frontmatter';
+import { type SkillLoadIssue, issue, loadSkillDirectory } from './skillLoader';
+import type { Skill } from './SkillSchema';
 
 // Local imports - utilities
 import type { Dirent } from 'node:fs';
 
-export type SkillIssueSeverity = 'error' | 'warning';
-
-export type SkillIssueCode =
-  | 'duplicate_name'
-  | 'duplicate_realpath'
-  | 'invalid_frontmatter'
-  | 'invalid_name'
-  | 'invalid_source'
-  | 'missing_source'
-  | 'missing_description'
-  | 'name_mismatch'
-  | 'read_error'
-  | 'source_read_error';
-
-export interface SkillLoadIssue {
-  severity: SkillIssueSeverity;
-  code: SkillIssueCode;
-  message: string;
-  path?: string;
-  name?: string;
-}
+export {
+  type SkillIssueCode,
+  type SkillIssueSeverity,
+  type SkillLoadIssue,
+} from './skillLoader';
 
 export interface DiscoverSkillsResult {
   skills: Skill[];
@@ -75,198 +49,6 @@ export interface SourcedSkill {
 export interface DiscoverSkillSourcesResult {
   skills: SourcedSkill[];
   errors: SkillLoadIssue[];
-}
-
-interface LoadedSkill {
-  skill?: Skill;
-  errors: SkillLoadIssue[];
-}
-
-function issue(
-  severity: SkillIssueSeverity,
-  code: SkillIssueCode,
-  message: string,
-  options: { path?: string; name?: string } = {},
-): SkillLoadIssue {
-  return { severity, code, message, ...options };
-}
-
-function firstZodMessage(error: ZodError): string {
-  return error.issues[0]?.message ?? 'Invalid skill metadata';
-}
-
-/**
- * `extractFrontmatter` throws plain errors whose messages begin with `SKILL.md`
- * or `Invalid SKILL.md`. Treat those as malformed frontmatter; anything else
- * (e.g. a failed `readFile`) is a read error.
- */
-function skillReadErrorCode(err: unknown): SkillIssueCode {
-  if (
-    err instanceof Error &&
-    (err.message.startsWith('SKILL.md') ||
-      err.message.startsWith('Invalid SKILL.md'))
-  ) {
-    return 'invalid_frontmatter';
-  }
-  return 'read_error';
-}
-
-function normalizeSkillName(
-  frontmatter: Record<string, unknown>,
-  directoryName: string,
-  skillPath: string,
-): { name?: string; errors: SkillLoadIssue[] } {
-  const errors: SkillLoadIssue[] = [];
-  const rawName = frontmatter.name;
-  const candidate = typeof rawName === 'string' ? rawName : directoryName;
-  const parsed = SkillNameSchema.safeParse(candidate);
-
-  if (parsed.success) {
-    const name = parsed.data;
-    if (typeof rawName === 'string' && name !== directoryName) {
-      errors.push(
-        issue(
-          'warning',
-          'name_mismatch',
-          `Skill name "${name}" does not match directory "${directoryName}"`,
-          { path: skillPath, name },
-        ),
-      );
-    }
-    return { name, errors };
-  }
-
-  if (rawName !== undefined) {
-    errors.push(
-      issue(
-        'warning',
-        'invalid_name',
-        `Ignoring invalid skill name: ${firstZodMessage(parsed.error)}`,
-        { path: skillPath },
-      ),
-    );
-  }
-
-  const fallback = SkillNameSchema.safeParse(directoryName);
-  if (fallback.success) {
-    return { name: fallback.data, errors };
-  }
-
-  errors.push(
-    issue(
-      'error',
-      'invalid_name',
-      `Directory name cannot be used as a skill name: ${firstZodMessage(
-        fallback.error,
-      )}`,
-      { path: skillPath },
-    ),
-  );
-  return { errors };
-}
-
-function normalizeSkillDescription(
-  frontmatter: Record<string, unknown>,
-  skillPath: string,
-  name: string,
-): { description?: string; errors: SkillLoadIssue[] } {
-  const errors: SkillLoadIssue[] = [];
-  const rawDescription = frontmatter.description;
-  const description =
-    typeof rawDescription === 'string'
-      ? collapseWhitespace(rawDescription)
-      : '';
-
-  if (!description) {
-    errors.push(
-      issue('error', 'missing_description', 'Skill description is required', {
-        path: skillPath,
-        name,
-      }),
-    );
-    return { errors };
-  }
-
-  if (description.length > SKILL_DESCRIPTION_MAX_LENGTH) {
-    errors.push(
-      issue(
-        'warning',
-        'invalid_frontmatter',
-        `Skill description exceeds ${SKILL_DESCRIPTION_MAX_LENGTH} characters and was truncated`,
-        { path: skillPath, name },
-      ),
-    );
-    return {
-      description: description.slice(0, SKILL_DESCRIPTION_MAX_LENGTH),
-      errors,
-    };
-  }
-
-  return { description, errors };
-}
-
-async function loadSkillDirectory(
-  skillDir: string,
-  directoryName: string,
-): Promise<LoadedSkill> {
-  const skillPath = path.join(skillDir, 'SKILL.md');
-
-  try {
-    const content = await fs.readFile(skillPath, 'utf8');
-    const { frontmatter, body } = extractFrontmatter(content);
-    if (!isObject(frontmatter)) {
-      return {
-        errors: [
-          issue(
-            'error',
-            'invalid_frontmatter',
-            'SKILL.md frontmatter must be a YAML object',
-            { path: skillPath },
-          ),
-        ],
-      };
-    }
-
-    const nameResult = normalizeSkillName(
-      frontmatter,
-      directoryName,
-      skillPath,
-    );
-    const errors = [...nameResult.errors];
-    if (!nameResult.name) return { errors };
-
-    const descriptionResult = normalizeSkillDescription(
-      frontmatter,
-      skillPath,
-      nameResult.name,
-    );
-    errors.push(...descriptionResult.errors);
-    if (!descriptionResult.description) return { errors };
-
-    const normalizedFrontmatter = {
-      ...frontmatter,
-      name: nameResult.name,
-      description: descriptionResult.description,
-    };
-    const skill = SkillSchema.parse({
-      name: nameResult.name,
-      description: descriptionResult.description,
-      body,
-      baseDir: skillDir,
-      path: skillPath,
-      frontmatter: normalizedFrontmatter,
-    });
-
-    return { skill, errors };
-  } catch (err) {
-    return {
-      errors: [
-        issue('error', skillReadErrorCode(err), toErrorMessage(err), {
-          path: skillPath,
-        }),
-      ],
-    };
-  }
 }
 
 /**
@@ -355,6 +137,48 @@ export async function discoverSkills(
 }
 
 /**
+ * Validate a `required` skill source, returning an issue when the path is
+ * missing or not a directory. Optional sources skip this check entirely.
+ */
+async function validateRequiredSource(
+  source: SkillSource,
+): Promise<SkillLoadIssue | undefined> {
+  try {
+    const sourceStat = await fs.stat(source.path);
+    if (!sourceStat.isDirectory()) {
+      return issue(
+        'error',
+        'invalid_source',
+        'Skill source is not a directory',
+        {
+          path: source.path,
+        },
+      );
+    }
+    return undefined;
+  } catch (err) {
+    if (isFileNotFoundError(err)) {
+      return issue('error', 'missing_source', 'Skill source does not exist', {
+        path: source.path,
+      });
+    }
+    if (isNotADirectoryError(err)) {
+      return issue(
+        'error',
+        'invalid_source',
+        'Skill source is not a directory',
+        {
+          path: source.path,
+        },
+      );
+    }
+    return issue('error', 'source_read_error', toErrorMessage(err), {
+      path: source.path,
+    });
+  }
+}
+
+/**
  * Discover skills from several roots in precedence order.
  *
  * The one-root loader remains useful for tests and direct imports. This wrapper
@@ -371,48 +195,9 @@ export async function discoverSkillSources(
 
   for (const source of sources) {
     if (source.required === true) {
-      try {
-        const sourceStat = await fs.stat(source.path);
-        if (!sourceStat.isDirectory()) {
-          errors.push(
-            issue(
-              'error',
-              'invalid_source',
-              'Skill source is not a directory',
-              {
-                path: source.path,
-              },
-            ),
-          );
-          continue;
-        }
-      } catch (err) {
-        if (isFileNotFoundError(err)) {
-          errors.push(
-            issue('error', 'missing_source', 'Skill source does not exist', {
-              path: source.path,
-            }),
-          );
-          continue;
-        }
-        if (isNotADirectoryError(err)) {
-          errors.push(
-            issue(
-              'error',
-              'invalid_source',
-              'Skill source is not a directory',
-              {
-                path: source.path,
-              },
-            ),
-          );
-          continue;
-        }
-        errors.push(
-          issue('error', 'source_read_error', toErrorMessage(err), {
-            path: source.path,
-          }),
-        );
+      const sourceError = await validateRequiredSource(source);
+      if (sourceError) {
+        errors.push(sourceError);
         continue;
       }
     }
